+++ pubhubs_hub/modules/pubhubs/__init__.py
@@ -1,2 +1,3 @@
 from .YiviRoomJoiner import YiviRoomJoiner
 from .DBMigration import DBMigration
+from .config_checker import ConfigChecker